Authorization on operation in data level implemented

+++ Restaurant.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -4,12 +4,14 @@ using Microsoft.Extensions.Logging;
 using Restaurants.Application.Restaurants.Commands.CreateRestaurant;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantRepository restaurantRepository) : IRequestHandler<DeleteRestaurantCommand>
+    IRestaurantRepository restaurantRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,7 +19,12 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         var restaurant = await restaurantRepository.GetByIdAsync(request.Id)
             ?? throw new NotFoundException<Restaurant>(request.Id.ToString());
 
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))
+            throw new ForbidException<Restaurant>(restaurant.Id.ToString());
+
         await restaurantRepository.DeleteAsync(restaurant);
+
+
     }
